transport_tcp: Treat EINTR and ERESTARTSYS as errors and do not retry.
Those are sent when the user runs drbdsetup disconnect. In
that case we do not wait for incoming connection. This is
a temporary patch only, because we should fix the signal
semantics (signals are SynchronizationEvents instead of
Notification events. We don't change that now since that
probably would break other things in the code.

On disconnect you will find error -4 messages in the
kernel log which you can ignore.