drbd: Fix race aborting connections claiming wrong protocol magic
The peer may process our P_CONNECTION_FEATURES and send a P_PING packet
using the magic for the agreed protocol version. P_PING is sent on the
control socket, so it may arrive before P_CONNECTION_FEATURES from the
peer, which is sent on the data socket.

I am fixing this by allowing any packet header version on the control
connection.

Co-developed-by: Joel Colledge <joel.colledge@linbit.com>