WinDRBD: Have RCU flags in drbd_resource (for begin/end state change).
In WinDRBD calls to rcu_read_lock() and rcu_read_unlock() may change
the IRQL. In order to be able to restore the original IRQL the
RCU calls were modified to return the old IRQL (rcu_read_lock)
and take the old IRQL as parameter (rcu_read_unlock()). Usually
this is inserted automatically by the 001-rcu_read_lock.cocci
coccinelle (spatch) script. But this works only if both calls
are in the same function.

This manual patch adds a windrbd_rcu_flags member to struct
drbd_resource and uses this storage for the RCU calls in
begin_state_change and end_state_change. It is assumed that
begin_state_change and end_state_change are called pairwise
and cannot recurse.