WinDRBD: have initializer function for transport_classes_lock
On Windows in order to initialize a semaphore one has to
call KeInitializeSemaphore(). On Linux this can be done
statically by the DECLARE_RWSEM() macro. This approach
does not work in Windows since we cannot call the function
implicitly at driver load.

Instead there is a init_tranport() function which has to
be called by the DriverEntry function of WinDRBD.