drbd_transport: disable sk_state_change during free
during kernel socket shutdown we might see state changes that lead to
NetworkFailure states. This seems to be a rare race, that is more likely
to happen on TLS-enabled sockets, perhaps because they have a longer
"shutdown sequence".

To fix, we reset the sk_state_change pointer for the control stream
when freeing the sockets.