drbd: Optimize AL updates in a two-node cluster
In the two-node case, a Secondary may safely omit activity log updates
if the Primary peer can write its activity log. This can improve
latency unless storage transaction latency is negligible compared to
replication link latency.

In any configuration with more than two nodes with a backing disk, the
activity log is necessary, as it might be needed if we recover from a
common power failure and the former primary is missing when the power
comes back.