drbd: fix compatibility with older peers in checksum resync
Older peers may send meaningless data in the block_id field of
P_RS_CANCEL and other "neg ack" packets. This occurs in the checksum
resync because the block_id field of struct drbd_peer_request was a
union with the digest pointer. The block_id was not reset in the "neg
ack" cases, so the digest pointer was sent on the wire instead.

This fixes errors of the form:
Unexpected resync reply at 53248s+524288