drbd: do not disconnect when sending resync request failed
When the connection is lost we want to transition to Unconnected and
then to Connecting. This line caused us to sometimes transition instead
to Disconnecting and then StandAlone.

Disconnect only if find_current_dagtag() returns an error.

Do not free the peer request here. It will be cleared up by
free_waiting_resync_requests(). This simplifies the code.