drbd: wait for rs_pending_cnt to clear on attach
This is necessary since the recent commit
c27694322e51 drbd: free peer req when disk fails or is detached during resync

Since that commit, if a resync peer request is pending when we detach,
that peer request will be removed when the reply arrives. We must not
reset rs_pending_cnt while the request is pending. Doing so leads to
an underflow when the reply is received.

Fix the case where we attach again. Instead of resetting the counter,
wait for there to be no resync requests pending.