drbd: make timeout during promotion attempt interruptible
When schedule_timeout_interruptible() ends early due to a signal, it
returns the remaining time, which is positive. So we should check for a
non-zero return value to detect this situation, rather than checking
whether the return value is negative.

schedule_timeout_interruptible() is a thin wrapper around
schedule_timeout(). From its docs:
> In all cases the return value is guaranteed to be non-negative.

Fixes: ac5bf0997080 ("drbd: Improve the try_to_promote retry loop")