drbd: Fix a subtle bug in set_cache_ptr_if_null()
The problem is that requests following a read request have the same
dagtag. As the code was, it moved the cache pointer to point to the
last of such requests having the same dagtag.

It's not an issue in a write-only workload.

One way to observe this bug is when the connection breaks and multiple
read requests are on the transfer log.

Then, DRBD will only re-issue the most recent of these read requests
to another peer.