drbd: send P_RS_CANCEL to peer that expects it
Even when that peer is now diskless.

Normally compatibility is maintained by making the newer version
compatible with the older version. This fix is the other way around. The
equivalent fix on the other side would be extremely complicated. This
fix is simple.

Without this fix, the peer gets stuck when attaching after being
diskless for a time. It repeatedly prints the message:
wait for rs_pending_cnt to clear