drbd: Fix a corner case leading to a hanging resync
In a cluster of A, B, C with A diskless and primary where the
connection from A to C is missing or down

   B
  / \
 A   C

And a resync running from B to C. Then the resync from B to C failed
to terminate if A executed a write after C sent its final resync
request to B and before B answered the last resync requests.

The diskless primary node (A) processes write requests by sending them
to the two diskful peers. One after another. From time to time, it
decides to send a P_PEER_ACK packet. With that, it intends to inform
the diskfull nodes who else got the write as well.

drbd_req_destroy()
  drbd_queue_peer_ack()
    w_queue_peer_ack()
      process_peer_ack_list()
        drbd_send_peer_ack()

When B receives that packet, it informs C that this block just became
out-of-sync and it needs to be resynced.

got_peer_ack()
  notify_sync_targets_or_free()
    drbd_queue_work()
      w_send_out_of_sync()
        drbd_send_out_of_sync()

C dutifully sets it out-of-sync. That is fine. It becomes only a
problem if C has already stopped scheduling, it's work for making
resync requests.

receive_out_of_sync()
  drbd_set_out_of_sync()

I am fixing that by restarting the resync_timer in this case.