compat: gate tls_get_record_type behind tls support
We already tied the tls_get_record_type patch to the general tls support.
However, we forgot to account for the case of a kernel having the
"tls_get_record_type" prototype without having TLS enabled.

Turns out that some kernel are configured that way, which lead to our patch
system complaining that no patch covers that case.

Turns out it is much simpler to gate the patch already in the gen_patch_names
tool.