WinDRBD: spin_lock_irq -> spin_lock_irqsave
Since rcu_read_lock changes the IRQL it is not safe to use
spin_lock_irq any more. We always have to check the flag.

There is a generic cocci script for that. But some cases
cannot be handled in cocci (like inter function dependencies)
therefore this manual patch exists.