drbd: Avoid wrong outdates in far_away_change()
C is diskless primary, and B is Inconsistent since A resyncs B. A is
UpToDate.

 A-->B   C

C connects with B. A should stay up-to-date (not become outdated).

 A-->B---C

The first part of the fix is that a primary node only declares other
nodes as weakly connected when it is a primary node with access to
data.

The transaction initiator then sets the TWOPC_PRI_INCAPABLE flag on
the commit packet of the two-phase state change when the primary nodes
do not declare other nodes as weakly connected.

A far-away node only outdates itself if it learns about distant
primaries and the two-phase commit packet does not have the
TWO_PC_PRI_INCAPABLE flag.