build: add minimal sanity check for test cases
To not end up with a completely bogus compat.h,
check that the compiler is "behaving" as expected,
and that the environment looks basically sane.

Add two test cases that should always succeed respectively always fail,
and refuse to even test-compile the other test cases
if this expectation is not met.