compat: build handshake.ko, if our test case says so
In Kbuild, we check against kernel version,
in the drbd-kernel-compat/tests/, we check for functionality.

On older distro kernels with backported TLS, this can result in
drbd_transport_tcp using symbols that would be provided by the handshake module,
but our Kbuild would, based on kernel version, decide to not build that module.
We end up with an unloadable transport.

We can bring the test in sync with our Kbuild, and add a check on kernel
version to drbd/drbd-kernel-compat/tests/have_tls_tx_rx.c, to disable
TLS for all older kernel versions, ignoring potentially backported
functionality:
`#if !IS_ENABLED(CONFIG_TLS) || LINUX_VERSION_CODE < KERNEL_VERSION(4,17,0)`

Or, now that the real Kbuild runs only after we generated our compat.h,
we can bring our Kbuild in sync with our test.
Which is what I try to do here.