WinDRBD: Ensure that only one fencing handler is running at the same time
When using long running fencing scripts (> 30 seconds or so) it
may happen that the fencing script is called several times when
a connection is half opened while the original fencing script is
running. This may lead to lots of races and confusion in the
fencing scripts.

This patch checks if a fencing script is already running and if
yes waits for it to terminate. Then the usual processing continues
(like exiting fencing I/O suspension mode if the script stonithed
the peer). For these actions then return value of the fencing script
started earlier is used.

Also on connection unregister threads waiting for a fencing script
terminating are terminated (with an invalid fencing script return
value since the fencing script is still running).