drbd: Allow a resync operation if sync source not connected with primary
Allow a resync to proceed even if the application write it depends on
originating from a currently unconnected node. Here is an example that
illustrates the case:

B and C have a backing disk, and A is diskless and primary.

    A
   / \
  B---C

Then the connection A-C gets interrupted. With that, C becomes
Outdated.  A continues writing data.

    A
   /
  B---C

Now, the connection A-B dies. B starts a resync to C.

    A

  B-->C

While the resync is ongoing, A connects to C.

    A
     \
  B-->C

Now C refers to node A with its resync requests' depend_dagtag_node_id.

This case is alredy handled by check_resync_source(). We should
not delay the resync from the resync source side.