compat: detect removal of sendpage directly
Upstream removed the ->sendpage() ops on the socket in favour of using sendmsg.
We used the newly introduced MSG_SPLICE_PAGES flag to detect this.

Turns out that the flag was backported to some LTS kernels (notably Debian 12).
The flag was backported without any of the supporting infrastructure, so some
implementations for sendmsg() reject the call based on this unknown flag.

The fix is to directly check the presence of the sendpage ops, as that is what
we are actually trying to patch.

Fixes: 6fd35546c2d2 ("compat: sock: Remove ->sendpage*() in favour of sendmsg(MSG_SPLICE_PAGES)")