build: fix builds when no compat patches are necessary
Recent changes to the build system meant that when the list of patches to
apply is empty, the build would fail because the rendered patch file did
not exist.

So we first check of the rendered patch file actually exists. If not, we
simply put an empty file in it place.

Also, use tabs for the gen_compat_patch.sh script consistently for the
sanity of anyone touching it.