Only outdate on disconnect when there is currently a connection.
When the connection fails and there is a long running fence-peer
handler and someone disconnects the resource on the Primary side
fencing (no I/O) mode is exited. This may lead to split brain
situations.

This patch adds an additional check if there is a valid connection
and outdates the peer disk only if we are currently connected to
the peer.