drbd: Do not clog the own worker thread for an empty 2PC
The DRBD resource work queue is not made for long-running work
items. Also, a lot of effort was necessary to avoid deadlocks when
work items need random DRBD locks. The final nail in the coffin of
this approach is two-phase commits that run into a 30-second
timeout. It is not acceptable that other work items get delayed for 30
seconds.
Remove all that. Use a system work queue instead.