drbd: fix cleanup of remote state changes upon connection cleanup
The idea of considering whether this state change targets the local
node was wrong to begin with. The question is whether the two-phase
commit can continue. Consider:

 A---B
  \ /
   C

When a two-phase commit from A addresses all nodes (with that
implicitly also B), and B loses the connection to A, it does not need
to abort it if it has C as a parent of the two-phase commit.

Remove the special case for no single connection up. Instead, check if
a remote two-phase commit still has one parent and check if a commit
is possible with all nested (or child) nodes.