docker: fix rpm-provides parsing
The rpm command shipped with RHEL7 requires the "-p" option to read from
uninstall RPM files. Without this option, the output was empty, which
lead to the comparison with the kernel.provides file during RPM selection
to report a match for the first RPM file it could find.

To fix this, use the "-p" option when querying the RPM file, and do a
sanity check that the result is actually non-empty.