drbd: fix potential circular deadlock in down/del-connection
If a drbdsetup (down, del-connection) is in del_connection() for peer B,
while concurrently an other connection C "fails",
or is closed from that peer, it can happen that

* drbdsetup in del_connection() for B and tries to drbd_flush_workqueue(),
waiting for the worker thread to make progress,

* the worker thread is doing some "after state change work"
and currently tries to drbd_notify_peers_lost_primary(),
waiting for the receiver thread of the lost peer C to "exit its mainloop",

* and the receiver for C is waiting for the conf_update mutex
before calling drbd_transport_shutdown().

Result is circular deadlock in drbdsetup, drbd worker and drbd receiver.

As long as some work may depend on the conf_mutex,
even if indirectly as in this case,
we must not hold the conf_mutex when calling drbd_flush_workqueue().

It should be sufficient to hold conf_mutex only directly around
drbd_unregister_connection()