drbd: fix potential page fault for "page not present"
Fixes: e88fdafcc (drbd: when scanning the bitmap, use the inline find_next_bit variant, 2023-09-12)

find_next_bit() and find_next_zero_bit() expect an (unsigned long *),
and will dereference it.  When scanning our bitmap, we are interested in 32bit
words of it.  The "current 32 bit word pointer" may point to the last 32 bits
in a page.  For 64bit long, if the page after the current page is not mapped,
this causes "page fault - not-present page".

Duplicate the "fast path" of these functions,
simplified for our use case of size:32, offset:0.