drbd: allocate the bio with allocating the peer_request
Instead of allocating the bio when submitting the peer_request,
allocate the bio when allocating the peer_request.
This is a preparation step for removing the page_chain from the
peer_request. We invented the drbd_page_chain with the idea that we
only have one chunk of data within a 4096 byte page.
That was a valid assumption back then. To move forward to a zero-copy
receive scheme, we need to eliminate these drbd_page_chains.