drbd: print node IDs when starting threads
It is useful to have the node IDs explicitly in the logs. This saves us
from having to infer them from other messages.

Stop printing the current task info. The reasons why we did this are
lost in the mists of time. It would make more sense to print the task
info for the new thread. Perhaps that was the original intention and it
has simply been broken for the last 20 years.

Move the drbd_thread_start(worker) call to after set_resource_options()
so that resource->res_opts.node_id is set when we need it. It is more
robust to start the thread only once everything else is initialized
anyway.