build: spatch result depends on gen_patch_names.c
The result of spatch, our "compat.patch", obviously depends on the
selection of (cocci) patches to apply, which not only depends on the
test results, but also on the tool we use to make the choices.

Add gen_patch_names.c to the dependencies, so both regular "make" as
well as "make spatch" (explicitly ask for a rebuild of the currently
relevant compat.patch) will pick up on changes in the tool and re-run spatch.