compat: check for disabled CONFIG_NET_HANDSHAKE
We need to apply the compat for the tls_get_record_type function if the
kernel has the symbol in the headers, but no actual handshake module. This
happens on new kernels (>=6.6) where the handshake module is disabled.

In this case, we do build our own handshake module for compatibility, but
we forgot to patch the necessary helper functions that are also part of the
handshake module.