drbd: manually resolving split-brain sometimes needs a full sync
If we notice data divergence ("split brain") because we find an identical data
generation "somewhere in the history uuids", we do no longer have a suitable
bitmap to use.

The "auto resolve" path already handles this by forcing a full sync.

This also force a full sync if it is manually resolved via "--discard-my-data".

Before, in this scenario we would end up with a bitmap based resync,
based on a not suitable bitmap. In general that means even if you follow the
documented procedure to resolve this kind of "split brain", data would still
not be identical on both nodes -- but reported as such.
That would be a possible source of data corruption later.

Note: "Normal" split brains are not affected (identical bitmap uuids,
we can safely use a bitmap based resync).
But after complex "multiple failure scenarios", this can be relevant.