rdma: Do not in block dtr_free() for excessive time (minutes)
Waiting in dtr_free() until the RDMA code did the final callbacks
sometimes blocked `drbdsetup del-peer` or `drbdsetup down` too long
(minutes). It is much more reasonable to have quick down and del-peer
operations and to keep only a reference to the module until the RDMA code
finishes its cleanup operations.

Also, if a low-level RDMA/Infiniband driver fails to clean up, people
no longer blame DRBD because one of the DRBD resources became stuck.
Here is an example stack trace:

[<0>] dtr_free+0x232/0x270 [drbd_transport_rdma]
[<0>] drbd_transport_shutdown+0x65/0xb0 [drbd]
[<0>] drbd_unregister_connection+0x122/0x230 [drbd]
[<0>] del_connection+0xb1/0x1a0 [drbd]
[<0>] adm_disconnect+0x1c7/0x270 [drbd]
[<0>] drbd_adm_del_peer+0x13/0x20 [drbd]
[<0>] genl_family_rcv_msg_doit+0xe7/0x150
[<0>] genl_rcv_msg+0xe2/0x1f0
[<0>] netlink_rcv_skb+0x56/0x100
[<0>] genl_rcv+0x29/0x40
[<0>] netlink_unicast+0x223/0x340
[<0>] netlink_sendmsg+0x24b/0x4c0
[<0>] sock_sendmsg+0x69/0x70
[<0>] sock_write_iter+0x93/0xf0
[<0>] new_sync_write+0x190/0x1a0
[<0>] vfs_write+0x1d5/0x270
[<0>] ksys_write+0xb5/0xf0
[<0>] __x64_sys_write+0x19/0x20
[<0>] do_syscall_64+0x5c/0xc0
[<0>] entry_SYSCALL_64_after_hwframe+0x61/0xcb

Because drbdsetup holds the adm_mutex at this point, it was no longer
possible to alter its configuration or to take it down.
With the new code, the damage is only an excess reference on the
drbd_transport_rdma module.