drbd: fix exposed_uuid going backward
When a diskless node in the primary role creates a new current UUID,
it sends it all its UpToDate peers and sets it as its own exposed
UUID. If there is only one UpToDate node in the secondary role and it
sends its current UUID to the diskless primary at about the same time,
then with the following state change, through the function
ensure_expose_data_uuid(), the diskless primary sets its exposed_uuid
back to the previous one.

Avoid this problem by not accepting a current_uuid update from a peer
when I am primary (authoritative), and I have told the peer the latest
update, and that peer is already on that latest (authoritative) value.