WinDRBD: Call WinDRBD when becoming Primary/Secondary.
Since there is no Windows device (disk) while secondary we
create it when becoming Primary and destroy it when becoming
Secondary. The reason behind that is that disk devices are
accessed (at least read) by Windows even when they are offline.
So it makes more sense to not have a Windows device than just
returning I/O errors.