drbd: Fix initialization of connect_state
Initializing peer_device->connect_state in conn_connect2() is too
late. This node might receive the initial UUIDs and the initial state
before it manages to call into conn_connect2().

If that happens connect_state.conn (drbd_repl_state) is set to 0, and
in the logs this appears as a questionmark. Like this line:
pdsk( Outdated -> DUnknown ) repl( Off -> ? ) [connected]

Fix all that by initializing everything for a connect two-phase-commit
with the new function drbd_init_connect_state().