drbd: re-order "drbd_fsync_device" in last close
We do no longer distinguish between read-only openers and writeable
openers.  The last one to close is responsible for sync'ing any caches
do the device.  This is actually handled by the generic block layer
already, but we still want the side effects of confirming that all the
data is on all the peers, even if this is asynchronous replication.

This should be an elaborate no-op, but clarifies that all code paths
dealing with write requests still see an open_cnt > 0.

It should also make it easier to correctly use the "exact open count"
for older kernels that still pass the "mode" to _open() and _release().