rdma: fix free() of scheduled delayed_work
dtr_cm objects can survive much longer than a DRBD connection. In
dtr_free() we ensure that the path->cs.retry_connect_work delayed_work
is no longer scheduled.
But when an RDMA_CM_EVENT_UNREACHABLE comes in after all that, we call
dtr_cma_disconnect(), which schedules the mentioned delayed_work.
At the end of the dtr_cma_disconnect_work_fn(), the final ref on the
cm gets dropped, which drops the final reference on the path.
With that the delayed_work was freed, since it is embedded in the
dtr_path.

Not schedule that delayed_work when the active_state is PCS_INACTIVE.