rdma: fix an access after free() in dtr_destroy_cm()
dtr_cm objects can survive much longer than a DRBD connection. We can
not assume in dtr_destroy_cm() that the transport still exists.
Accessing cm_count in the rdma_transport is a access after free().

Remove the cm_count and, with that, the access to it in dtr_destroy_cm().