drbd_transport: fix a use after free in drbd_get_listener()
In the moment the listener is exposed on the path (after releasing the
spinlock), a drbd_put_listener() might remove it from the path and put
one reference. If that was the last reference, the listener gets
destroyed.
Avoid that by exposing a listener on the path after its initialization
is completed.