rdma: fix concurrency of activate_path() and __dtr_disconnect_path()
Let's assume dtr_activate_pth() executes past the atomic_set()
statements and sets the passive_state to PCS_CONNECTING.

Then, maybe on another CPU, __ctr_disconnect_path() runs. It does the
atomic_cmpchx(), sets the passive_state back to PCS_INACTIVE, and
calls drbd_put_listener(). The listener is not yet there, so it is a
no-op.

Then execution continues in dtr_activate_path(), and it missed that it
should no longer listen for incoming connections.

Fix that race condition by re-checking the passive state after
creating the listener and putting it in case it is no longer needed.