drbd: Fix drbd ending in inconsistent repl states after connect
The node with the smaller node-id always initiates the two-phase
commit of a connect operation. It is the active node. The node with
the higher node ID passively waits to receive the P_TWOPC_PREPARE
packet.
During the handshake sequence, the active node receives the initial
state packet from the passive node. The peer might interrupt the
active node by aborting the two-phase commit. If another two-phase
commit was already waiting, that gets executed next. That might happen
while its receiver thread is still processing packets from the aborted
two-pc.
If that happens, it prematurely declares the second connect two-phase
commit as completed, resulting in a repl_state of Established when it
should be WFBitMapS.

There is already code to avoid this problem, which is waiting until
the receiver processes the initial state of the previous
two-phase-commit before beginning the new one. It is the call to
wait_initial_states_receive() in change_cluster_wide_state(). The bug
was that the INITIAL_STATE_RECEIVED bit was still set when the
two-phase commit was executed for the second time.

Clearing INITIAL_STATE_RECEIVED before repeating the connect two-phase
commit fixes this.