drbd: do drbd_notify_peers_lost_primary() from conn_disconnect()
Instead of making the worker thread wait for the receiver of the lost
peer to terminate and then notifying the peers from the worker, just
notify them directly from the receiver before it terminates.

This fixes a complicated deadlock which can occur as follows:

* Peer S is Secondary, we are SyncTarget towards S.
* Peer P is Primary.
* We send a resync request to S.
* P sends a write to us which is held due to a conflict with the resync
  request.
* We lose connection to P, our receiver for P waits in conn_disconnect()
  for active_ee_cnt to reach zero. That is, for the write to complete.
* Our worker calls drbd_notify_peers_lost_primary() and waits for the
  receiver for P to terminate.
* S sends P_STATE, perhaps because it also lost connection to P. Our
  receiver for S waits in receive_state() for the worker to process a
  state change.
* S responds to the resync request, but we never receive the response
  because our receiver for S is stuck.

The deadlock dependency cycle in summary:
Write from P
-> Resync request to S
-> receiver for S
-> worker
-> receiver for P
-> (back to the top)

Break this cycle by removing the dependency from the worker to the
receiver.