drbd: Improve pretty-print of cluster-wide state change
The union/struct drbd_state dates far back, probably to the days of
drbd-8.0. It has a single field .conn for the connection state. Later
on, the meaning of that field was extended in that it contains the
replication state on an established connection.  In most parts of the
code, that is quite natural; it is the replication state when the
struct/packet contains that drbd_state word; it is the connection
state when the container is about a connection.

The drbd_print_cluster_wide_state_change() function is very generic
(its first argument is a resource), and it lacks the context
information if the drbd_state is about a connection or a peer_device.

So far, the called drbd_conn_str() when .conn was about a
repl_state. The result was a question mark `?` in the log when one
starts a full-sync using the invalidate or invalidate-remote commands.

Fix that by calling drbd_repl_str() based on the value.