drbd: only do reconciliation resync if our disk is Consistent
This prevents a bug which causes us to get stuck in
WFBitMapT/Established. The bug occurred when we have no connection to a
Primary node, but a peer does. When that peer loses the connection to
the Primary, it sends us P_PEER_DAGTAG. Then we may try to start a
reconciliation resync. But we don't send P_PEER_DAGTAG, so the peer does
not start its side of the reconciliation resync. We end up stuck.

This situation only occurs when the peer is on the same current UUID as
us. That is, there were no writes on the Primary since it was promoted.
If there were writes then we enter a normal resync before P_PEER_DAGTAG
is received and the reconciliation resync is skipped due to the
replication state.

Fix this bug by ignoring P_PEER_DAGTAG if our disk is already Outdated.
In fact, we only expect to do a reconciliation resync when our disk is
Consistent, so test for precisely that case.