Unplug paranoia: use a positive value.
In the dagtag_newer_eq macro the values are cast to
signed (s64) for comparision. If we add (1<<63) to a
small integer we get a negative value when cast to
s64. Therefore the dagtag_newer_eq in the 'paranoia'
(see comment) case does the opposite, namely causing
endless loop of unplug events being sent. You can check
this by looking at the CPU load it would show a CPU
at 100%.

This patch uses a still large but not too large value
to add in the paranoia case.