drbd: fix wait_event() in receive_peer_dagtag()
We need to send P_PEER_DAGTAG before we start the resync. Otherwise,
the peer is surprised to get a bitmap, and it will complain about
being in the wrong state for receiving a bitmap.

So we must wait before starting the resync until the receiver thread
to the lost primary did that.

Fixes: c1bcb0c4c ("drbd: do drbd_notify_peers_lost_primary() from conn_disconnect()")