drbd: do multiple resync passes instead of skipping back
This is preparation for disabling replication during resync.

We only detect the end of the resync as SyncTarget from
make_resync_request(). This is scheduled regardless of whether there are
any more requests to make and is used to poll for the conditions that
are necessary to finish the resync. As a result, this change may cause
us to finish the resync slightly later. Finishing the resync immediately
once the requests complete can be added back later.

One consequence of this is that resyncs can last longer. Suppose we have
a chain A-B-C with A Primary, a resync from B to C, and continuous write
activity on A. In this situation C will collect collect out-of-sync bits
while waiting to notice that the resync pass is finished. Then it will
start another pass due to the out-of-sync bits. This can occur
arbitrarily often.