drbd: An incapable primary should not outdate a far UpToDate secondary
Let's consider a diskless node A with a resync from C to B. B is
Primary, and C is up-to-date.

   A
  / \
 B<--C

Then the resync is interrupted

   A
  / \
 B   C

B becomes a primary without access to UpToDate data. It either throws
I/O errors or it freezes I/O. In either case, it should not outdate C.

Fix this by detecting B as an incapable primary when evaluating the
data gathered in the empty two-phase commit after losing connection to
a primary.