drbd_transport_lb-tcp: Fix counting connected_paths
drbd_path_established() returns true when the path is established. It
does not distinguish between whether it was already marked as
established or just got established. Therefore, its return code can
not be used to increase the connected_path count.

Fix that by incrementing the counter only if we set the TR_ESTABLISHED
bit (= it was not set before).