docker: find RPM provides from symvers directly
Using the RPM macros to generate the symbols provided by the currently running
kernel takes a lot of time: every module is manually unpacked, read and
processed. However, the same information is already available in the
"/lib/modules/*/symvers*" file. We just have to parse it to generate
output compatible with the existing infrastructure.

In a quick test, this reduces the load time for the DRBD module from minutes
to one second.