drbd: flush requests from before resync before finishing
Otherwise there might be requests in the pipeline which are sent as
P_OUT_OF_SYNC packets to the SyncTarget node. For instance, when the
replication state is WFBitMapS, P_OUT_OF_SYNC packets are sent. If the
resync finishes before these packets arrive, the blocks are not synced.
This can result in silent data corruption.

2-node examples
===============

The above situation can occur when an empty resync is triggered. An
empty resync can be triggered, for instance, by the use of the
new-current-uuid command while the nodes are not connected.

It may also be possible to trigger the issue with a non-empty resync due
to the following race.

1. A write on Primary is submitted while it is WFBitMapS to peer. Before
   the sender processes it...
2. A resync request is sent and handled on the SyncSource up to the
   point of the w_e_end_rsdata_req work item being executed.
3. The SyncTarget receives the resync reply and finishes the resync.
4. ...and then sender on the Primary processes the write and sends a
   P_OUT_OF_SYNC packet.

This is unlikely to occur because it requires a backing device read to
complete before the sender has processed a request that has already been
submitted.

3-node example
==============

A similar situation with 3 or more nodes is much more likely. Suppose we
have nodes: P diskless Primary, S UpToDate, T Outdated:

1. P-S connected
2. P-T connected
3. S-T connect, become SyncSource-SyncTarget
4. T sends Inconsistent state to P, before that arrives...
5. Resync runs and finishes
6. Write on P sent to S, not T
7. ...P receives T's Inconsistent state

In this case, the blocks would normally be added to the resync when P
sends P_PEER_ACK to S and S sends P_OUT_OF_SYNC to T. The problems occur
as above when the resync finishes before all this happens.

Fix
===

Fix these problems by introducing a flush mechanism.

When a node becomes SyncTarget, it sends P_FLUSH_REQUESTS to each
Primary peer. Before finishing the resync, the SyncTarget waits for the
corresponding P_FLUSH_REQUESTS_ACK from each UpToDate peer. This
indicates that all P_OUT_OF_SYNC from before the start of the resync
have been transmitted.

When a Primary receives P_FLUSH_REQUESTS, it stores where it is in the
change stream. When all requests up to this point have been processed
and the corresponding P_PEER_ACK packets have been sent, it sends
P_FLUSH_REQUESTS_ACK back to the SyncTarget and P_FLUSH_FORWARD to the
other peers.

When a node receives P_FLUSH_FORWARD, it ensures that all P_OUT_OF_SYNC
for previously received P_PEER_ACK packets have been sent, and then
sends P_FLUSH_REQUESTS_ACK to the SyncTarget initiator of the flush.

The result is that the SyncTarget only finishes the resync after it is
made aware of all out-of-sync bits set for requests before the resync
started.